Moe H3 handling to separate file
