+++ picohttp/picohttp.vcxproj.filters
@@ -19,6 +19,9 @@
     <ClCompile Include="webtransport.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="h3zero_server.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="democlient.h">
